+++ src/units/http-kernel/index.js
@@ -1,5 +1,5 @@
 // import factory server
-import factoryServer from '../../support/http/factory'
+import factoryServer from '../../support/http/factory-server'
 
 // import core unit.
 import { middleware } from '../core'
@@ -7,6 +7,7 @@ import { middleware } from '../core'
 // import route definitions
 import routeDefinitions from './route-definitions'
 
+// create http kernel
 const httpKernel = factoryServer(middleware, routeDefinitions)
 
 // default router export.
